+++ tests/timeseries/test_resampling.py
@@ -407,7 +407,7 @@ async def test_future_samples_not_included(
         a_sequence(sample0s, sample1s), config, source_props  # sample2_1s is not here
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=3, sampling_period_s=None
+        sampling_start=timestamp, received_samples=3, sampling_period=None
     )
     assert _get_buffer_len(resampler, source_receiver) == config.initial_buffer_len
     sink_mock.reset_mock()
@@ -489,7 +489,7 @@ async def test_resampling_with_one_window(
         a_sequence(sample1s), config, source_props
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=2, sampling_period_s=None
+        sampling_start=timestamp, received_samples=2, sampling_period=None
     )
     assert _get_buffer_len(resampler, source_recvr) == config.initial_buffer_len
     sink_mock.reset_mock()
@@ -518,7 +518,9 @@ async def test_resampling_with_one_window(
     # By now we have a full buffer (5 samples and a buffer of length 4), which
     # we received in 4 seconds, so we have an input period of 0.8s.
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=5, sampling_period_s=0.8
+        sampling_start=timestamp,
+        received_samples=5,
+        sampling_period=timedelta(seconds=0.8),
     )
     # The buffer should be able to hold 2 seconds of data, and data is coming
     # every 0.8 seconds, so we should be able to store 3 samples.
@@ -536,7 +538,9 @@ async def test_resampling_with_one_window(
         current_iteration=3,
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=5, sampling_period_s=0.8
+        sampling_start=timestamp,
+        received_samples=5,
+        sampling_period=timedelta(seconds=0.8),
     )
     assert _get_buffer_len(resampler, source_recvr) == 3
 
@@ -598,7 +602,7 @@ async def test_resampling_with_one_and_a_half_windows(  # pylint: disable=too-ma
         a_sequence(sample0s, sample1s), config, source_props
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=2, sampling_period_s=None
+        sampling_start=timestamp, received_samples=2, sampling_period=None
     )
     assert _get_buffer_len(resampler, source_recvr) == config.initial_buffer_len
     sink_mock.reset_mock()
@@ -626,7 +630,7 @@ async def test_resampling_with_one_and_a_half_windows(  # pylint: disable=too-ma
         a_sequence(sample2_5s, sample3s, sample4s), config, source_props
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=5, sampling_period_s=None
+        sampling_start=timestamp, received_samples=5, sampling_period=None
     )
     assert _get_buffer_len(resampler, source_recvr) == config.initial_buffer_len
     sink_mock.reset_mock()
@@ -654,7 +658,9 @@ async def test_resampling_with_one_and_a_half_windows(  # pylint: disable=too-ma
     # By now we have a full buffer (7 samples and a buffer of length 6), which
     # we received in 4 seconds, so we have an input period of 6/7s.
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=7, sampling_period_s=6 / 7
+        sampling_start=timestamp,
+        received_samples=7,
+        sampling_period=timedelta(seconds=6 / 7),
     )
     # The buffer should be able to hold 2 * 1.5 (3) seconds of data, and data
     # is coming every 6/7 seconds (~0.857s), so we should be able to store
@@ -693,7 +699,9 @@ async def test_resampling_with_one_and_a_half_windows(  # pylint: disable=too-ma
         current_iteration=5,
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=7, sampling_period_s=6 / 7
+        sampling_start=timestamp,
+        received_samples=7,
+        sampling_period=timedelta(seconds=6 / 7),
     )
     assert _get_buffer_len(resampler, source_recvr) == 4
 
@@ -755,7 +763,7 @@ async def test_resampling_with_two_windows(  # pylint: disable=too-many-statemen
         a_sequence(sample0s, sample1s), config, source_props
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=2, sampling_period_s=None
+        sampling_start=timestamp, received_samples=2, sampling_period=None
     )
     assert _get_buffer_len(resampler, source_recvr) == config.initial_buffer_len
     sink_mock.reset_mock()
@@ -783,7 +791,7 @@ async def test_resampling_with_two_windows(  # pylint: disable=too-many-statemen
         a_sequence(sample1s, sample2_5s, sample3s, sample4s), config, source_props
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=5, sampling_period_s=None
+        sampling_start=timestamp, received_samples=5, sampling_period=None
     )
     assert _get_buffer_len(resampler, source_recvr) == config.initial_buffer_len
     sink_mock.reset_mock()
@@ -811,7 +819,7 @@ async def test_resampling_with_two_windows(  # pylint: disable=too-many-statemen
         source_props,
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=7, sampling_period_s=None
+        sampling_start=timestamp, received_samples=7, sampling_period=None
     )
     assert _get_buffer_len(resampler, source_recvr) == config.initial_buffer_len
     sink_mock.reset_mock()
@@ -833,7 +841,7 @@ async def test_resampling_with_two_windows(  # pylint: disable=too-many-statemen
         a_sequence(sample5s, sample6s), config, source_props
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=7, sampling_period_s=None
+        sampling_start=timestamp, received_samples=7, sampling_period=None
     )
     assert _get_buffer_len(resampler, source_recvr) == config.initial_buffer_len
     sink_mock.reset_mock()
@@ -849,7 +857,7 @@ async def test_resampling_with_two_windows(  # pylint: disable=too-many-statemen
         current_iteration=5,
     )
     assert source_props == SourceProperties(
-        sampling_start=timestamp, received_samples=7, sampling_period_s=None
+        sampling_start=timestamp, received_samples=7, sampling_period=None
     )
     assert _get_buffer_len(resampler, source_recvr) == config.initial_buffer_len